# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/desenvolvedores": {
            "get": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Listar desenvolvedores",
                "operationId": "59ee2c434a8f547891b5d76ab82e9757",
                "responses": {
                    "200": {
                        "description": "Lista de desenvolvedores"
                    },
                    "404": {
                        "description": "Nenhum desenvolvedor encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Criar desenvolvedor",
                "operationId": "79d206e8107cf767aacd707922a55048",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nivel_id",
                                    "nome",
                                    "sexo",
                                    "data_nascimento",
                                    "hobby"
                                ],
                                "properties": {
                                    "nivel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "sexo": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "hobby": {
                                        "type": "string",
                                        "example": "Programação"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Desenvolvedor criado"
                    }
                }
            }
        },
        "/api/desenvolvedores/paginate": {
            "get": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Listar desenvolvedores com paginação",
                "operationId": "bef63b859b4cce34548a8d10e6bff892",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Tamanho da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de desenvolvedores paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nivel": {
                                                        "type": "string"
                                                    },
                                                    "nome": {
                                                        "type": "string"
                                                    },
                                                    "sexo": {
                                                        "type": "string"
                                                    },
                                                    "data_nascimento": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "idade": {
                                                        "type": "integer"
                                                    },
                                                    "hobby": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum desenvolvedor encontrado"
                    }
                }
            }
        },
        "/api/desenvolvedores/{id}": {
            "get": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Mostrar desenvolvedor",
                "operationId": "0316a08cbe3c5a219cd11204dd95e1fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do desenvolvedor"
                    },
                    "404": {
                        "description": "Desenvolvedor não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Atualizar desenvolvedor",
                "operationId": "f19a4f4f281c735290a0cd2e43516916",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nivel_id",
                                    "nome",
                                    "sexo",
                                    "data_nascimento",
                                    "hobby"
                                ],
                                "properties": {
                                    "nivel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "sexo": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "hobby": {
                                        "type": "string",
                                        "example": "Programação"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Desenvolvedor atualizado"
                    },
                    "404": {
                        "description": "Desenvolvedor não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Deletar desenvolvedor",
                "operationId": "ab714e67b8b76ccf2e0f7799ae557b43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Desenvolvedor deletado"
                    },
                    "404": {
                        "description": "Desenvolvedor não encontrado"
                    }
                }
            }
        },
        "/api/niveis": {
            "get": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Listar níveis",
                "operationId": "f983113f46d3765dfcacc06fe03b483c",
                "responses": {
                    "200": {
                        "description": "Lista de níveis"
                    },
                    "404": {
                        "description": "Nenhum nível encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Criar nível",
                "operationId": "372073b181193e713a4c2631caf170f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nivel"
                                ],
                                "properties": {
                                    "nivel": {
                                        "type": "string",
                                        "example": "Junior"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nível criado"
                    }
                }
            }
        },
        "/api/niveis/paginate": {
            "get": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Listar níveis com paginação",
                "operationId": "fab77f4f54d1172d94fcce21e0d2771f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Tamanho da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de níveis paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nivel": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum nível encontrado"
                    }
                }
            }
        },
        "/api/niveis/{id}": {
            "get": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Mostrar nível",
                "operationId": "40938244cfd89332e1a939aa93c953c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do nível"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Atualizar nível",
                "operationId": "360c2793e59e977ef778fed3f550babf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nivel"
                                ],
                                "properties": {
                                    "nivel": {
                                        "type": "string",
                                        "example": "Pleno"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nível atualizado"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Deletar nível",
                "operationId": "95b63514dd90008c4185702e4c5f9ea9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nível deletado"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Desenvolvedores",
            "description": "API Endpoints for Desenvolvedores"
        },
        {
            "name": "Níveis",
            "description": "API Endpoints for Níveis"
        }
    ]
}